update integer types to bigint

+++ db/migrate/20211006130827_create_coins.rb
@@ -6,20 +6,20 @@ class CreateCoins < ActiveRecord::Migration[6.1]
       t.string :name
       t.string :image
       t.float :current_price
-      t.integer :market_cap
-      t.integer :market_cap_rank
-      t.integer :fully_diluted_valuation
-      t.integer :total_volume
-      t.integer :high_24h
-      t.integer :low_24h
+      t.bigint :market_cap
+      t.bigint :market_cap_rank
+      t.bigint :fully_diluted_valuation
+      t.bigint :total_volume
+      t.bigint :high_24h
+      t.bigint :low_24h
       t.float :price_change_24h
       t.float :price_change_percentage_24h
-      t.integer :market_cap_change_24h
+      t.bigint :market_cap_change_24h
       t.float :market_cap_change_percentage_24h
-      t.integer :circulating_supply
-      t.integer :total_supply
-      t.integer :max_supply
-      t.integer :ath
+      t.bigint :circulating_supply
+      t.bigint :total_supply
+      t.bigint :max_supply
+      t.bigint :ath
       t.float :ath_change_percentage
       t.datetime :ath_date
       t.float :atl
